Significant refactor of command handling
This commit standardizes the return values of command executions. By
returning a Dictionary<string, object> instead of a string, it is now
possible to perform pre-processing of the execution result before
returning, and post-processing to add sessionId after returning.
Furthermore, it makes it much easier to determine which type of HTTP
response to send to the client.